Combine multiple createApis to one
- This will bring performance boost and also we can "resetApi" when we log out
- We need to reset all api-calls when we are logged out and showing the landing page
