+++ src/baseApi.ts
@@ -1,0 +1,9 @@
+import { createApi } from '@reduxjs/toolkit/query/react';
+import { refreshingBaseQuery } from '@/utils/http';
+
+export const baseApi = createApi({
+  baseQuery: refreshingBaseQuery,
+  reducerPath: 'api',
+  tagTypes: ['mentors', 'myuser'],
+  endpoints: () => ({}),
+});
