+++ src/features/Authentication/useAuthenticated.ts
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 import { useAppDispatch, useAppSelector } from '@/store';
-import { selectIsLoggedIn } from '@/features/Authentication/userSlice';
+import { selectIsLoggedIn } from '@/features/Authentication/selectors';
 import { authenticationApi } from '@/features/Authentication/authenticationApi';
 
 const useAuthenticated = (): boolean => {
