+++ src/features/Authentication/userSlice.ts
@@ -1,16 +1,10 @@
-import { createSelector } from 'reselect';
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
-import {
-  authenticationApi,
-  defaultAppUser,
-  defaultMentor,
-} from './authenticationApi';
-import { selectHasBeenChatting } from '../Chat/selectors';
+import { defaultAppUser } from './models';
+import { authenticationApi } from './authenticationApi';
 
-import type { Account, AppUser, User } from './authenticationApi';
-import type { ApiMentor } from '../MentorPage/mentorPageApi';
-import type { RootState } from '@/store';
+import type { Account, AppUser, User } from './models';
+import type { ApiMentor } from '../MentorPage/models';
 
 const initialState: AppUser = defaultAppUser;
 
@@ -41,41 +35,5 @@ export const user = createSlice({
       });
   },
 });
-
-export const selectAccount = ({ user }: RootState) => user.account;
-
-export const selectMentor = ({ user }: RootState) =>
-  user.mentor ?? defaultMentor;
-
-export const selectUser = ({ user }: RootState) => user.user;
-
-export const selectIsLoggedIn = createSelector(selectUser, ({ id }) =>
-  Boolean(id),
-);
-
-export const selectAccountId = createSelector(selectAccount, ({ id }) => id);
-
-export const selectUserId = createSelector(selectUser, ({ id }) => id);
-
-export const selectUserRole = createSelector(selectUser, ({ role }) => role);
-
-export const selectIsMentee = createSelector(
-  selectUserRole,
-  role => role === 'mentee',
-);
-
-export const selectIsMentor = createSelector(
-  selectUserRole,
-  role => role === 'mentor',
-);
-
-export const selectAppRole = createSelector(
-  selectUserRole,
-  selectHasBeenChatting,
-  (userRole, hasBeenChatting) => {
-    if (userRole === 'mentee' && !hasBeenChatting) return 'freshMentee';
-    return userRole;
-  },
-);
 
 export const { setAccount, setMentor, setUser } = user.actions;
