+++ src/features/MentorPage/components/MentorList/MentorCard/Expanded/Header.tsx
@@ -1,4 +1,4 @@
-import type { Mentor } from '@/features/MentorPage/mentorPageApi';
+import type { Mentor } from '@/features/MentorPage/models';
 
 import { getStatus } from '@/utils/utils';
 import { useGetLayoutMode } from '@/hooks/useGetLayoutMode';
