+++ src/features/MentorPage/components/MentorList/MentorCard/Expanded/index.tsx
@@ -1,9 +1,9 @@
-import type { Mentor } from '@/features/MentorPage/mentorPageApi';
+import type { Mentor } from '@/features/MentorPage/models';
 
 import { useGetLayoutMode } from '@/hooks/useGetLayoutMode';
 import { useEscape } from '@/hooks/useEscape';
 import { useAppSelector } from '@/store';
-import { selectUserId } from '@/features/Authentication/userSlice';
+import { selectUserId } from '@/features/Authentication/selectors';
 import { getIsOlderThanDaysAgo } from '@/utils/utils';
 
 import styled, { css } from 'styled-components';
