+++ src/features/MentorPage/components/MentorList/MentorCard/List/ExpandButton.tsx
@@ -1,4 +1,4 @@
-import type { Mentor } from '@/features/MentorPage/mentorPageApi';
+import type { Mentor } from '@/features/MentorPage/models';
 
 import { useTranslation } from 'react-i18next';
